Allow InstallSpinnaker.sh to ommit cassandra (#1026)
When omitting cassandra, it will still add the package manager location
but will not actually install it. I can remove that too, but figured I'd
wait until we completely remove cassandra. In the meantime adding the location
will make it easier to manually install cassandra later should the deployer change
their mind.

The default behavior for GCP is to leave out cassandra. The other platforms are unchanged
for the time being. Adding a --noinstall_cassandra will force it off and --install_cassandra
will force it on.

Removed explicit cassandra (and redis) management entirely from developer mode (running from source).
To use cassandra from source, you'll need to ensure it is running yourself.
If it is an upstart service (as is normal) then it will already be running anyway.

I left the stop scripts in for now because I'm not sure if anyone is using the stop dependencies
script, but expect to remove this as well.

+++ pylib/spinnaker/change_cassandra.py
@@ -62,6 +62,16 @@ _FRONT50_KEYS = ['front50.cassandra.enabled', 'front50.redis.enabled',
 SPINNAKER_INSTALLED_PATH = '/opt/spinnaker/cassandra/SPINNAKER_INSTALLED_CASSANDRA'
 SPINNAKER_DISABLED_PATH = '/opt/spinnaker/cassandra/SPINNAKER_DISABLED_CASSANDRA'
 
+
+def cassandra_installed():
+  return os.system('service --status-all 2>1 | grep -q cassandra') == 0
+
+
+def service_is_running(name):
+  return os.system('service --status-all 2>1 | grep -q {name} | grep " + "'
+                   .format(name=name)) == 0
+
+
 class CassandraChanger(object):
   @classmethod
   def init_argument_parser(cls, parser):
@@ -110,25 +120,32 @@ class CassandraChanger(object):
     if os.path.exists(SPINNAKER_INSTALLED_PATH):
       os.remove(SPINNAKER_INSTALLED_PATH)
       open(SPINNAKER_DISABLED_PATH, 'w').close()
-    with open('/etc/init/cassandra.override', 'w') as f:
+
+    if cassandra_installed():
+      with open('/etc/init/cassandra.override', 'w') as f:
         f.write('manual')
-    print 'Stopping cassandra service...'
-    os.system('service cassandra stop')
+      print 'Stopping cassandra service...'
+      os.system('service cassandra stop || true')
 
   def enable_cassandra(self):
+    is_installed = cassandra_installed()
     if os.path.exists(SPINNAKER_DISABLED_PATH):
       os.remove(SPINNAKER_DISABLED_PATH)
       open(SPINNAKER_INSTALLED_PATH, 'w').close()
-    try:
-      os.remove('/etc/init/cassandra.override')
-    except OSError as err:
-      if err.errno == errno.ENOENT:
-         pass
+
+    if is_installed:
+      try:
+        os.remove('/etc/init/cassandra.override')
+      except OSError as err:
+        if err.errno == errno.ENOENT:
+           pass
 
     cassandra_host = self.__configurator.bindings['services.cassandra.host']
     cassandra_port = self.__configurator.bindings['services.cassandra.port']
     if (cassandra_host in ['localhost', '127.0.0.1', '0.0.0.0']
         or cassandra_host == socket.gethostname()):
+      if not is_installed:
+        raise RuntimeError('Cassandra is not installed - install cassandra and try again.')
       print 'Starting cassandra service...'
       os.system('service cassandra start')
       sock = socket.socket()
@@ -190,20 +207,24 @@ class CassandraChanger(object):
     self.maybe_restart_spinnaker()
 
   def maybe_restart_spinnaker(self):
-    if not self.__configurator.bindings.get(
+    if not service_is_running('echo'):
+      print 'Echo wasnt running'
+    elif not self.__configurator.bindings.get(
           'services.echo.{which}.enabled'.format(which=self.__options.echo)):
       print 'Restarting echo...'
-      os.system('service echo restart')
+      os.system('service echo restart || true')
     else:
       print 'Echo was unchanged.'
 
-    if (not self.__configurator.bindings.get(
+    if not service_is_running('front50'):
+      print 'Front50 wasnt running'
+    elif (not self.__configurator.bindings.get(
            'services.front50.{which}.enabled'.format(which=self.__options.front50))
         or (self.__options.front50 in ['s3', 'gcs']
             and self.__configurator.bindings['services.front50.storage_bucket']
                != self.__options.bucket)):
       print 'Restarting front50...'
-      os.system('service front50 restart')
+      os.system('service front50 restart || true')
     else:
       print 'Front50 was unchanged.'
